requirements.txt.txt
# ECG Analysis System - Requirements
# Python 3.8+

# Deep Learning
torch>=2.0.0
torchvision>=0.15.0

# Data Processing
numpy>=1.24.0
pandas>=2.0.0
scikit-learn>=1.3.0

# Image Processing
Pillow>=10.0.0
opencv-python>=4.8.0

# Signal Processing
scipy>=1.11.0

# Visualization
matplotlib>=3.7.0

# PDF Generation
fpdf>=1.7.2

# Serial Communication (for ESP32)
pyserial>=3.5

# Additional utilities
tqdm>=4.65.0